requirements.txt
fastapi
uvicorn
pydantic-ai
nest-asyncio
qdrant-client
ollama
weaviate-client
langchain
langchain-openai
langchain-community 
langchain_ollama
langchain-weaviate
langchain-qdrant
pypdf
